+++ src/components/toolbox/EditTextBox.tsx
@@ -87,7 +87,6 @@ const EditTextBox = ({
 
     return getPointPositionBeforeCanvasTransformation(
       [borders.x, borders.y],
-      [-shape.translation[0], -shape.translation[1]],
       shape.rotation,
       center
     )
@@ -107,7 +106,8 @@ const EditTextBox = ({
       data-fontbold={shape.style?.fontBold ?? false}
       data-fontitalic={shape.style?.fontItalic ?? false}
       contentEditable={!disabled}
-      onInput={updateContentEditable}></StyledEditBox>
+      onInput={updateContentEditable}
+    />
   )
 }