+++ src/utils/intersect.ts
@@ -28,35 +28,6 @@ export const getCursorPosition = (
   ]
 }
 
-export const getSettingsPosition = (
-  shape: DrawableShape,
-  canvas: HTMLCanvasElement | null,
-  givenWidth: number,
-  givenHeight: number,
-  selectionPadding: number
-) => {
-  const shapeInfos = getShapeInfos(shape, selectionPadding)
-
-  const positionInCanvas = getPointPositionBeforeCanvasTransformation(
-    shapeInfos.center,
-    [-shape.translation[0], -shape.translation[1]],
-    shape.rotation,
-    shapeInfos.center
-  )
-
-  const canvasBoundingRect = canvas?.getBoundingClientRect() ?? {
-    left: 0,
-    top: 0,
-    width: givenWidth,
-    height: givenHeight
-  }
-
-  return [
-    canvasBoundingRect.left + positionInCanvas[0] * (canvasBoundingRect.width / givenWidth),
-    canvasBoundingRect.top + positionInCanvas[1] * (canvasBoundingRect.height / givenHeight)
-  ]
-}
-
 const isPointInsideRect = (rect: Rect, point: Point) => {
   return (
     point[0] >= rect.x &&
@@ -69,12 +40,11 @@ const isPointInsideRect = (rect: Rect, point: Point) => {
 export const getPointPositionAfterCanvasTransformation = (
   [positionX, positionY]: Point,
   [canvasOffsetX, canvasOffsetY]: Point,
-  [translationX, translationY]: Point,
   shapeRotation: number,
   [centerX, centerY]: Point
 ): Point => {
-  const newX = positionX + canvasOffsetX - centerX - translationX
-  const newY = positionY + canvasOffsetY - centerY - translationY
+  const newX = positionX + canvasOffsetX - centerX
+  const newY = positionY + canvasOffsetY - centerY
   const rotatedY = newY * Math.cos(shapeRotation) - newX * Math.sin(shapeRotation)
   const rotatedX = newY * Math.sin(shapeRotation) + newX * Math.cos(shapeRotation)
 
@@ -85,7 +55,6 @@ export const getPointPositionAfterCanvasTransformation = (
 
 export const getPointPositionBeforeCanvasTransformation = (
   [positionX, positionY]: Point,
-  [translationX, translationY]: Point,
   shapeRotation: number,
   [centerX, centerY]: Point
 ): Point => {
@@ -96,7 +65,7 @@ export const getPointPositionBeforeCanvasTransformation = (
 
   const translatedX = rotatedX + centerX
   const translatedY = rotatedY + centerY
-  return [translatedX - translationX, translatedY - translationY]
+  return [translatedX, translatedY]
 }
 
 export const applyRotationToVector = (vector: Point, shapeRotation: number): Point => {
@@ -122,7 +91,6 @@ export const checkPositionIntersection = (
   const newPosition = getPointPositionAfterCanvasTransformation(
     position,
     canvasOffset,
-    shape.translation,
     shape.rotation,
     center
   )